Сгрупировал утилиты и константы
