+++ src/mock/filters-mock.js
@@ -1,4 +1,4 @@
-import { filter } from '../constant/filter-const.js';
+import { filter } from '../utils-constant/constant.js';
 //console.log(filter);
 const generateFilters = (points) => Object.entries(filter).map(([filterType, filterPatternByType]) => ({
   type: filterType,
