+++ src/view/filter-view.js
@@ -1,4 +1,4 @@
-import { FiltersPoint } from '../constant/filter-const.js';
+import { FiltersPoint } from '../utils-constant/constant.js';
 import AbstractView from '../framework/view/abstract-view';
 import { capitalizeFirstLetter } from '../utils-constant/utils.js';
